+++ lib/intl_phone_picker/view_helpers.rb
@@ -1,16 +1,14 @@
 module IntlPhonePicker
   module ViewHelpers
 
-    def intl_phone_tag_fr(name, value = nil, options = {})
-      # options[:class] ||= String.new
-      # options[:class] += ' intl_phone_input_fr'
-      telephone_field_tag(name, value, options.merge(:class => 'intl_phone_input_fr'))
-    end
+    def intl_phone_tag(name, value = nil, options = {})
+      available_languages = ['us', 'fr']
 
-    def intl_phone_tag_us(name, value = nil, options = {})
-      # options[:class] ||= String.new
-      # options[:class] += ' intl_phone_input_us'
-      telephone_field_tag(name, value, options.merge(:class => 'intl_phone_input_fr'))
+      if (I18n.locale && I18n.locale.to_s.in?(available_languages))
+        telephone_field_tag(name, value, options.merge(:class => 'intl_phone_input_' + I18n.locale.to_s))
+      else
+        telephone_field_tag(name, value, options.merge(:class => 'intl_phone_input_us'))
+      end
     end
 
   end